Add files via upload
Wheat productivity as low as 25-30 mound/acre is a serious issue and likely to trigger the food security issue in Pakistan. Besides, many factor imbalance use of fertilizer is one of major culprit for such low yield. Our Integrated Nutr. Management aims at closing the yield gap so that the food security can be ensured.